AND-6774 Move Blockchain extensions to lib:blockchain_sdk
