+++ app/src/main/java/com/tangem/tap/domain/walletconnect2/app/TangemWcBlockchainHelper.kt
@@ -1,13 +1,13 @@
 package com.tangem.tap.domain.walletconnect2.app
 
 import com.tangem.blockchain.common.Blockchain
-import com.tangem.domain.common.extensions.fromNetworkId
-import com.tangem.domain.common.extensions.toNetworkId
+import com.tangem.blockchainsdk.utils.fromNetworkId
+import com.tangem.blockchainsdk.utils.toNetworkId
 import com.tangem.tap.domain.walletconnect2.domain.WcBlockchainHelper
 import com.tangem.tap.domain.walletconnect2.toggles.WalletConnectFeatureToggles
 
 internal class TangemWcBlockchainHelper(
-    private val featureToggles: WalletConnectFeatureToggles,
+    featureToggles: WalletConnectFeatureToggles,
 ) : WcBlockchainHelper {
 
     private val supportedNonEvmBlockchains = if (featureToggles.isSolanaTxSignEnabled) {
