+++ app/src/main/java/com/tangem/tap/features/customtoken/impl/data/DefaultCustomTokenRepository.kt
@@ -1,10 +1,10 @@
 package com.tangem.tap.features.customtoken.impl.data
 
 import com.tangem.blockchain.common.Blockchain
+import com.tangem.blockchainsdk.utils.toNetworkId
 import com.tangem.datasource.api.common.response.getOrThrow
 import com.tangem.datasource.api.tangemTech.TangemTechApi
 import com.tangem.domain.common.extensions.supportedBlockchains
-import com.tangem.domain.common.extensions.toNetworkId
 import com.tangem.domain.common.util.cardTypesResolver
 import com.tangem.tap.features.customtoken.impl.data.converters.FoundTokenConverter
 import com.tangem.tap.features.customtoken.impl.domain.CustomTokenRepository
