+++ app/src/main/java/com/tangem/tap/proxy/TransactionManagerImpl.kt
@@ -21,10 +21,10 @@ import com.tangem.blockchain.extensions.Result
 import com.tangem.blockchain.extensions.SimpleResult
 import com.tangem.blockchain.externallinkprovider.TxExploreState
 import com.tangem.blockchain.network.ResultChecker
+import com.tangem.blockchainsdk.utils.fromNetworkId
 import com.tangem.common.core.TangemSdkError
 import com.tangem.common.extensions.hexToBytes
 import com.tangem.domain.card.repository.CardSdkConfigRepository
-import com.tangem.domain.common.extensions.fromNetworkId
 import com.tangem.domain.walletmanager.WalletManagersFacade
 import com.tangem.lib.crypto.TransactionManager
 import com.tangem.lib.crypto.models.*
