+++ data/feedback/build.gradle.kts
@@ -37,6 +37,7 @@ dependencies {
 
     implementation(projects.domain.feedback)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.models)
     implementation(projects.domain.wallets.models)
 }
