+++ data/tokens/src/main/kotlin/com/tangem/data/tokens/paging/CoinsResponseConverter.kt
@@ -1,9 +1,9 @@
 package com.tangem.data.tokens.paging
 
 import com.tangem.blockchain.common.Blockchain
+import com.tangem.blockchainsdk.utils.fromNetworkId
 import com.tangem.data.tokens.utils.getNetworkStandardType
 import com.tangem.datasource.api.tangemTech.models.CoinsResponse
-import com.tangem.domain.common.extensions.fromNetworkId
 import com.tangem.domain.tokens.model.CryptoCurrency
 import com.tangem.domain.tokens.model.Token
 import com.tangem.utils.converter.Converter
