+++ data/transaction/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
     /** Domain */
     implementation(projects.domain.transaction)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.wallets.models)
     implementation(projects.domain.tokens.models)
