+++ data/txhistory/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
     implementation(projects.core.utils)
     implementation(projects.core.datasource)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.tokens.models)
     implementation(projects.domain.txhistory)
     implementation(projects.domain.txhistory.models)
