+++ data/visa/build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
     /** Project - Utils */
     implementation(projects.core.utils)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
 
     /** Project - Libs */
     debugImplementation(projects.libs.visa)
