+++ domain/card/build.gradle.kts
@@ -14,6 +14,7 @@ dependencies {
     implementation(projects.domain.demo)
     implementation(projects.domain.core)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     // TODO: Remove after new card scan result was implemented
     implementation(projects.domain.models)
     implementation(projects.domain.tokens.models)
