+++ domain/tokens/build.gradle.kts
@@ -14,6 +14,7 @@ dependencies {
     api(projects.domain.core)
     implementation(projects.domain.models)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.tokens.models)
     implementation(projects.domain.txhistory.models)
     implementation(projects.domain.wallets.models)
