+++ domain/transaction/build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
 
     implementation(projects.domain.models)
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.wallets.models)
     implementation(projects.domain.tokens)
     implementation(projects.domain.tokens.models)
