+++ domain/wallets/build.gradle.kts
@@ -17,6 +17,7 @@ dependencies {
 
     // region Domain modules
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.models)
     implementation(projects.domain.tokens)
     implementation(projects.domain.tokens.models)
