+++ features/referral/data/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
 
     /** Domain modules */
     implementation(projects.domain.legacy)
+    implementation(projects.libs.blockchainSdk)
     implementation(projects.domain.models)
     implementation(projects.domain.tokens.models)
     implementation(projects.domain.wallets.models)
